Avoid missing NFCIP inbound frame.
If a NFCIP frame is received (NXP_EVT_NFC_RCV_DATA event) before a call to
phHal4Nfc_Receive() function, the frame was lost. This could happen if the local
host acts as a Target and the remote peer sends its first frame too fast.

The data is actually saved in a temporary buffer but, when the phHal4Nfc_Receive()
is eventualy called, it triggers a callback with a status code
NFCSTATUS_MORE_INFORMATION, which is handled as an error.

This patch fixes this behaviour and makes sure that the correct status
code is provided with the frame itself.

Change-Id: I5311762ca70c71f98dea0287b0844eb1c8dddb6b