Fixed LLCP receive function when accessing RW.
In some situations, the working buffer may be full and some packets
are stored in the receive window when trying to read from upper
layer.

In this case, the receive function first reads data from the working
buffer and if possible, also tries to read data stored in the receive
window. The previous receive window data extraction algorithm could
lead to an infinite loop. With this fix, we can ensure that this won't
happen any more.

Change-Id: Iddb412213f4e9cb5fb42691bd282dbf0a21a936e