Add support for connecting to NfcA on IsoDep tags.
Always add a ISO14443-3A device handle for tags that are
found through the A-gate. This allows the upper stack
to connect and maintain a connection to the NfcA technology,
even if the tag has IsoDep on top of NfcA. The reason this
can work is that libnfc does do the IsoDep activation in
software, and if the NfcA target is on a separate handle
this activation will not take place if you connect to
this handle.

Bug: 4195403
Change-Id: I00015e74bdadc6666c2971f8102244670a76e340