Increase number of LLCP sockets from 5 to 10.
We currently have 3 server sockets open (snep, npp, bluetooth); I found
that when we traverse from snep to npp (as a fallback), then do
bluetooth, we run out of sockets (NFCSTATUS_INSUFFICIENT_RESOURCES
is returned on socket creation). This is because cleanup of the sockets
is asynchronous and the previous 2 sockets may still be disconnecting.
Increasing to 10 should be enough for ICS purposes.

Change-Id: I803da4cbcc97e7c5b736bf7276d2bbb7136c6dc3