Deal with LLCP send errors properly (part 1).
Send failures on LLCP sockets were not dealt with at all - the transport
was just left in a state where it thought there was a pending send
and a pending callback. Before we typically only did one send and
it didn't bite us so much, but now with multiple sockets we run
into this frequently.

This really just fixes part of the problem - we should also look
if state rollbacks are necessary, as those are not done either
on send failures.

Change-Id: I484ca598e043ead835251aedaabe47d3a1da7a8f