Use a 10s timeout in uart_read() in FW download mode.
This provides a path to flush RX, necessary when we enter FW download mode
with bogus data already in RX. This fixes a problem where NFC was dead on several
HSPA and LTE devices.

Also decrease FW timeout completition from 120s to 60s. Typically FW download
takes 30s, and 120 is unnecessarily large and causes a long delay to bring up
NFC when we need to retry FW.

Bug: 5468674
Change-Id: I0fd40c199daac861fe5acda4f2d214bdcff8e869