Use NFC HAL for some libnfc settings.
The NFC HAL in libhardware allows us to store and retrieve
product-specific settings. This patch moves two product-specific
settings to the HAL: the device port, and the usage of the i2c
workaround we had for crespo. This means configuring the port
no longer needs to be done from JNI land.

Change-Id: I2e19b6f188f808bc2f2a1f1abc28f2a6c47e6a4c