DO NOT MERGE
Disable NXP_SMX (secure element) in libnfc-nxp for Phantasm

Phantasm uses a PN544, which has no secure element.  libnfc-nxp is
configured to assume a secure element currently and doesn't handle
runtime detection properly.  The result is that on first boot,
when the NfcService tries to wipe the secure element, it hangs
as it tries to open a connection to the non-existing secure element.
Changing the define to NXP_SMX=0 makes the library report no
secure element, and the wipe operation fails (like it should) when
it tries to do the wipe.

This fixes

Bug: 5964890

Will file a separate bug to have the library and Nfc.apk do better
at runtime detection of the secure element.

This shouldn't go back to master since it's a target specific change.

Change-Id: Id6e57b7f5a88e96a59b9cdc6f67ac966fa964356
Signed-off-by: Mike J. Chen <mjchen@google.com>