Support of topaz tag NFC Forum version number 1.1
The latest version of NFC forum Tag type 1 spec is 1.1. This version
number needs to be supported by libnfc to support Topaz tags V1.1 with
dynamic content. Version number is part of capability container which
resides in segment 1 of EEPROM memory of the tag. For older versions of
Tag type 1, NFC device may be backward compatible.

Change-Id: Ibbfa720d2e34c2f2612579e368f31f4a5c66b465
Author: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 33236