Immediately bounce back SYMM packet when initiator.
On first establishment of the LLCP link, if we are
the initiator, we waited the full LTO before sending
the SYMM to the target. This will prevent the target
from communicating with us for 750ms! Old Android
implementations only connected LLCP on touch, but
newer ones will immediately connect LLCP.

To solve this, bounce the SYMM back immediately, to allow
the target to start sending as soon as it wants to.

Change-Id: I2fb09599798527f2e6f908b8ba362e0061120be2